Adjusting "minus()" and "divide()" function
Noticed that the current "a - b" should be reversed for the minus() and divide() function. Current minus() adds a "-" to the front of each answer which then throws each answer after that off if you continue to do anything else to the current number. The divide() function instead change the answer to a decimal (ie: 8/4).